bill-verification/requirements.txt
pandas==1.0.5
requests==2.22.0
Django==3.1.2
python_dateutil==2.8.1
